requirements.txt
streamlit>=1.28.0
together>=0.2.7
python-dotenv>=1.0.0
